Move all test scenes to visual namespace
