+++ osu.Game.Rulesets.Rush.Tests/Visual/TestSceneGround.cs
@@ -5,7 +5,7 @@ using osu.Framework.Graphics;
 using osu.Game.Rulesets.Rush.UI.Ground;
 using osu.Game.Tests.Visual;
 
-namespace osu.Game.Rulesets.Rush.Tests
+namespace osu.Game.Rulesets.Rush.Tests.Visual
 {
     public class TestSceneGround : OsuTestScene
     {
