+++ osu.Game.Rulesets.Rush.Tests/Visual/TestSceneRushAutoplay.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Shane Woolcock. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Rulesets.Rush.Tests
+namespace osu.Game.Rulesets.Rush.Tests.Visual
 {
     public class TestSceneRushAutoplay : TestSceneRushPlayer
     {
